+++ examples/heptagram.rs
@@ -1,7 +1,8 @@
 // heptagram.rs
-use footile::{FillRule, Path2D, Plotter, Transform};
+use footile::{FillRule, Path2D, Plotter};
 use pix::matte::Matte8;
 use pix::Raster;
+use pointy::Transform;
 
 mod png;